+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Improved
 - moved vendors to separate folder
+- added a option to disable tracking link showing in Woocommerce emails
+- the settings page is organized into sections
 
 ## [1.9.0]
 ### Fixed